requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.19
pandas==2.0.3
llama-cpp-python==0.1.103
